* tests/jffs2_[123].c: Fixed compiler warnings for parameter         mismatch for diag_printf().

+++ packages/fs/jffs2/current/tests/jffs2_3.c
@@ -68,8 +68,8 @@
 #define NELEM(_x_) (sizeof(_x_)/sizeof(*(_x_)))
 
 #define SHOW_RESULT( _fn, _res ) \
-diag_printf("FAIL: " #_fn "() returned %d %s\n", _res, \
-            _res<0?strerror(errno):"");
+diag_printf("FAIL: " #_fn "() returned %ld %s\n", _res, \
+            (unsigned long) _res<0?strerror(errno):"");
 
 //==========================================================================
 // file creation, deletion and testing functions
